Fix display issue regarding operation and operator
In the previous version we created a workaround to write a functional calculator.
After the user clicked an operator button, the program displayed 0.
This version we solved the issue and removed the workaround.
Now when the user clicks an operator or the equal button, the program displays the current number or the result until the user clicks another button.